refactor: ToprankingResponses 응답 객체 반환명 TopRankingResponse로 변경
